fedimint_api_client::api::global_api::with_cache

Struct GlobalFederationApiWithCache

pub struct GlobalFederationApiWithCache<T> {
    pub(crate) inner: T,
    pub(crate) await_session_lru: Arc<Mutex<LruCache<u64, Arc<OnceCell<SessionOutcome>>>>>,
    pub(crate) get_session_status_lru: Arc<Mutex<LruCache<u64, Arc<OnceCell<SessionOutcome>>>>>,
}

IGlobalFederationApi wrapping some T: IRawFederationApi and adding a tiny bit of caching.

Use GlobalFederationApiWithCacheExt::with_cache to create.

Fields
inner: Tawait_session_lru: Arc<Mutex<LruCache<u64, Arc<OnceCell<SessionOutcome>>>>>

Small LRU used as IGlobalFederationApi::await_block cache.

This is mostly to avoid multiple client module recovery processes re-requesting same blocks and putting burden on the federation.

The LRU can be be fairly small, as if the modules are (near-)bottlenecked on fetching blocks they will naturally synchronize, or split into a handful of groups. And if they are not, no LRU here is going to help them.

get_session_status_lru: Arc<Mutex<LruCache<u64, Arc<OnceCell<SessionOutcome>>>>>

Like Self::await_session_lru, but for IGlobalFederationApi::get_session_status.

In theory these two LRUs have the same content, but one is locked by potentially long-blocking operation, while the other non-blocking one. Given how tiny they are, it’s not worth complicating things to unify them.

Implementations

impl<T> GlobalFederationApiWithCache<T>

pub fn new(inner: T) -> GlobalFederationApiWithCache<T>

impl<T> GlobalFederationApiWithCache<T>
where T: IRawFederationApi + MaybeSend + MaybeSync + 'static,

pub(crate) async fn await_block_raw( &self, block_index: u64, decoders: &ModuleDecoderRegistry, ) -> Result<SessionOutcome>

pub(crate) fn select_peers_for_status( &self, ) -> impl Iterator<Item = PeerId> + '_

pub(crate) async fn get_session_status_raw_v2( &self, block_index: u64, broadcast_public_keys: &BTreeMap<PeerId, PublicKey>, decoders: &ModuleDecoderRegistry, ) -> Result<SessionStatus>

pub(crate) async fn get_session_status_raw( &self, block_index: u64, decoders: &ModuleDecoderRegistry, ) -> Result<SessionStatus>