fedimint_api_client::api::global_api::with_request_hook

Type Alias ApiRequestHook

pub type ApiRequestHook = Arc<dyn Fn(DynIRawFederationApi) -> DynIRawFederationApi + Send + Sync + 'static>;

“Api Request Hook”

An “api request hook” is a function that gets a raw federation api and can either pass it unmodified (no hook) or wrap it in whatever custom logic and return as a new raw federation api, possibly forwarding the call to the original one.

This is meant to allow downstream users to add custom logic for debugging, testing (e.g. simulating network being down), collecting stats, notifing about slow calls, errors, etc.

Aliased Type

struct ApiRequestHook { /* private fields */ }