Trait fedimint_api_client::api::IGlobalFederationApi

pub trait IGlobalFederationApi: IRawFederationApi {

    // Required methods
    fn submit_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx: Transaction
    ) -> Pin<Box<dyn Future<Output = FederationResult<SerdeModuleEncoding<Result<TransactionId, TransactionError>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn await_block<'life0, 'life1, 'async_trait>(
        &'life0 self,
        block_index: u64,
        decoders: &'life1 ModuleDecoderRegistry
    ) -> Pin<Box<dyn Future<Output = Result<SessionOutcome>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_session_status<'life0, 'life1, 'async_trait>(
        &'life0 self,
        block_index: u64,
        decoders: &'life1 ModuleDecoderRegistry
    ) -> Pin<Box<dyn Future<Output = Result<SessionStatus>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn session_count<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = FederationResult<u64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn await_transaction<'life0, 'async_trait>(
        &'life0 self,
        txid: TransactionId
    ) -> Pin<Box<dyn Future<Output = FederationResult<TransactionId>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn server_config_consensus_hash<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = FederationResult<Hash>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn upload_backup<'life0, 'life1, 'async_trait>(
        &'life0 self,
        request: &'life1 SignedBackupRequest
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn download_backup<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 PublicKey
    ) -> Pin<Box<dyn Future<Output = FederationResult<Vec<ClientBackupSnapshot>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn discover_api_version_set<'life0, 'life1, 'async_trait>(
        &'life0 self,
        client_versions: &'life1 SupportedApiVersionsSummary,
        timeout: Duration,
        num_responses_required: Option<usize>
    ) -> Pin<Box<dyn Future<Output = FederationResult<ApiVersionSet>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn set_password<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_config_gen_connections<'life0, 'async_trait>(
        &'life0 self,
        info: ConfigGenConnectionsRequest,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn add_config_gen_peer<'life0, 'async_trait>(
        &'life0 self,
        peer: PeerServerParams
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_config_gen_peers<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = FederationResult<Vec<PeerServerParams>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_default_config_gen_params<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<ConfigGenParamsRequest>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_config_gen_params<'life0, 'async_trait>(
        &'life0 self,
        requested: ConfigGenParamsRequest,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn consensus_config_gen_params<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = FederationResult<ConfigGenParamsResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn run_dkg<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_verify_config_hash<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<BTreeMap<PeerId, Hash>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn verified_configs<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<BTreeMap<PeerId, Hash>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn start_consensus<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn status<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = FederationResult<StatusResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn audit<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<AuditSummary>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn guardian_config_backup<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<GuardianConfigBackup>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn auth<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn restart_federation_setup<'life0, 'async_trait>(
        &'life0 self,
        auth: ApiAuth
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

The API for the global (non-module) endpoints

Required Methods

fn submit_transaction<'life0, 'async_trait>( &'life0 self, tx: Transaction ) -> Pin<Box<dyn Future<Output = FederationResult<SerdeModuleEncoding<Result<TransactionId, TransactionError>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn await_block<'life0, 'life1, 'async_trait>( &'life0 self, block_index: u64, decoders: &'life1 ModuleDecoderRegistry ) -> Pin<Box<dyn Future<Output = Result<SessionOutcome>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_session_status<'life0, 'life1, 'async_trait>( &'life0 self, block_index: u64, decoders: &'life1 ModuleDecoderRegistry ) -> Pin<Box<dyn Future<Output = Result<SessionStatus>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn session_count<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = FederationResult<u64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn await_transaction<'life0, 'async_trait>( &'life0 self, txid: TransactionId ) -> Pin<Box<dyn Future<Output = FederationResult<TransactionId>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn server_config_consensus_hash<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = FederationResult<Hash>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Fetches the server consensus hash if enough peers agree on it

fn upload_backup<'life0, 'life1, 'async_trait>( &'life0 self, request: &'life1 SignedBackupRequest ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn download_backup<'life0, 'life1, 'async_trait>( &'life0 self, id: &'life1 PublicKey ) -> Pin<Box<dyn Future<Output = FederationResult<Vec<ClientBackupSnapshot>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn discover_api_version_set<'life0, 'life1, 'async_trait>( &'life0 self, client_versions: &'life1 SupportedApiVersionsSummary, timeout: Duration, num_responses_required: Option<usize> ) -> Pin<Box<dyn Future<Output = FederationResult<ApiVersionSet>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Query peers and calculate optimal common api versions to use.

fn set_password<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Sets the password used to decrypt the configs and authenticate

Must be called first before any other calls to the API

fn set_config_gen_connections<'life0, 'async_trait>( &'life0 self, info: ConfigGenConnectionsRequest, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

During config gen, sets the server connection containing our endpoints

Optionally sends our server info to the config gen leader using add_config_gen_peer

fn add_config_gen_peer<'life0, 'async_trait>( &'life0 self, peer: PeerServerParams ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

During config gen, used for an API-to-API call that adds a peer’s server connection info to the leader.

Note this call will fail until the leader has their API running and has set_server_connections so clients should retry.

This call is not authenticated because it’s guardian-to-guardian

fn get_config_gen_peers<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = FederationResult<Vec<PeerServerParams>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

During config gen, gets all the server connections we’ve received from peers using add_config_gen_peer

Could be called on the leader, so it’s not authenticated

fn get_default_config_gen_params<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<ConfigGenParamsRequest>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the default config gen params which can be configured by the leader, gives them a template to modify

fn set_config_gen_params<'life0, 'async_trait>( &'life0 self, requested: ConfigGenParamsRequest, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Leader sets the consensus params, everyone sets the local params

After calling this ConfigGenParams can be created for DKG

fn consensus_config_gen_params<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = FederationResult<ConfigGenParamsResponse>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the consensus config gen params, followers will delegate this call to the leader. Once this endpoint returns successfully we can run DKG.

fn run_dkg<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Runs DKG, can only be called once after configs have been generated in get_consensus_config_gen_params. If DKG fails this returns a 500 error and config gen must be restarted.

fn get_verify_config_hash<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<BTreeMap<PeerId, Hash>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

After DKG, returns the hash of the consensus config tweaked with our id. We need to share this with all other peers to complete verification.

fn verified_configs<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<BTreeMap<PeerId, Hash>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Updates local state and notify leader that we have verified configs. This allows for a synchronization point, before we start consensus.

fn start_consensus<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Reads the configs from the disk, starts the consensus server, and shuts down the config gen API to start the Fedimint API

Clients may receive an error due to forced shutdown, should call the server_status to see if consensus has started.

fn status<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = FederationResult<StatusResponse>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the status of the server

fn audit<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<AuditSummary>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Show an audit across all modules

fn guardian_config_backup<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<GuardianConfigBackup>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Download the guardian config to back it up

fn auth<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Check auth credentials

fn restart_federation_setup<'life0, 'async_trait>( &'life0 self, auth: ApiAuth ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,