fedimint_api_client
Module api

Modules

error 🔒

global_api 🔒

net

peer 🔒

Structs

ApiVersionSet

Set of api versions for each component (core + modules)

DynGlobalApi

DynModuleApi

FederationError

An API request error when calling an entire federation

FederationStatus

The status of a server, including how it views its peers

GlobalFederationApiWithCache

IGlobalFederationApi wrapping some T: IRawFederationApi and adding a tiny bit of caching.

GuardianConfigBackup

Archive of all the guardian config files that can be used to recover a lost guardian node.

PeerResponse

PeerStatus

StatusResponse

WsFederationApi

Mint API client that will try to run queries against all peers expecting equal results from at least min_eq_results of them. Peers that return differing results are returned as a peer faults list.

Enums

JsonRpcClientError

Error type.

OutputOutcomeError

PeerConnectionStatus

PeerError

An API request error when calling a single federation peer

Traits

FederationApiExt

An extension trait allowing to making federation-wide API call on top IRawFederationApi.

GlobalFederationApiWithCacheExt

Convenience extension trait used for wrapping IRawFederationApi in a GlobalFederationApiWithCache

IGlobalFederationApi

The API for the global (non-module) endpoints

IModuleFederationApi

Trait marker for the module (non-global) endpoints

IRawFederationApi

An API (module or global) that can query a federation

JsonRpcClient

Functions

deserialize_outcome

Type Aliases

FederationResult

JsonRpcResult

OutputOutcomeResult

PeerResult

SerdeOutputOutcome