fedimint_api_client

Module api

Modules

error ๐

global_api

Structs

ApiVersionSet

Set of api versions for each component (core + modules)

DynGlobalApi

DynModuleApi

FederationApi

Federation API client

FederationError

An API request error when calling an entire federation

LegacyFederationStatus

The status of a server, including how it views its peers

LegacyPeerStatus

StatusResponse

Enums

LegacyP2PConnectionStatus

OutputOutcomeError

ServerError

An API request error when calling a single federation peer

Constants

VERSION_THAT_INTRODUCED_AWAIT_OUTPUTS_OUTCOMES

VERSION_THAT_INTRODUCED_GET_SESSION_STATUS

VERSION_THAT_INTRODUCED_GET_SESSION_STATUS_V2

Traits

FederationApiExt

An extension trait allowing to making federation-wide API call on top IRawFederationApi.

IGlobalFederationApi

The API for the global (non-module) endpoints

IModuleFederationApi

Trait marker for the module (non-global) endpoints

IRawFederationApi

An API (module or global) that can query a federation

Functions

deserialize_outcome

Type Aliases

FederationResult

OutputOutcomeResult

SerdeOutputOutcome

ServerResult