Trait fedimint_api_client::query::QueryStrategy

pub trait QueryStrategy<IR, OR = IR> {
    // Required method
    fn process(
        &mut self,
        peer_id: PeerId,
        response: PeerResult<IR>,
    ) -> QueryStep<OR>;
}

Fedimint query strategy

Due to federated security model each Fedimint client API call to the Federation might require a different way to process one or more required responses from the Federation members. This trait abstracts away the details of each specific strategy for the generic client Api code.

Required Methods

fn process( &mut self, peer_id: PeerId, response: PeerResult<IR>, ) -> QueryStep<OR>